Fixed name issue, added list of planets
Turns out the naming and planet type wasn't working, and had an
unecessary " - 1" which was preventing the last item in  the "name list"
and "planet type list" from being used.

In addition, planets are now being stored in the PLANETLIST LIST, which
allows for  planets to easily be browsed, sorted and etc.